+++ src/main/java/khodro/Khodro.java
@@ -1,0 +1,19 @@
+package khodro;
+
+public class Khodro  {
+
+    boolean ayaRoshanAst;
+
+    public void roshan() {
+        ayaRoshanAst = true;
+    }
+
+    void khamoush() {
+        ayaRoshanAst = false;
+    }
+
+    public boolean ayaDarHaleHarekatAst() {
+        return false;
+    }
+}
+
