+++ src/main/java/khodro/Motor.java
@@ -1,15 +1,7 @@
 package khodro;
 
-public class Motor {
-    boolean ayaRoshanAst;
+public class Motor extends Khodro {
 
-    public void roshan() {
-        ayaRoshanAst = true;
-    }
-
-    void khamoush() {
-        ayaRoshanAst = false;
-    }
 
     public boolean ayaDarHaleHarekatAst() {
         return ayaRoshanAst ;
